create matrix
added createMatrix() function, which finds all the values for the
MarkovMatrix from the data provided. The values in the matrix are
not normalized, but there is a vector normVec provided which holds
all normalization constant. This approach allows for more modular
use of the code. E.g.: live updating can easily be implemented